refactor(common): analytics를 common 레이어 밑으로 이동
