--- android/festago/app/src/main/java/com/festago/festago/analytics/AnalyticsHelper.kt
@@ -1,5 +1,0 @@
-package com.festago.festago.analytics
-
-interface AnalyticsHelper {
-    fun logEvent(event: AnalyticsEvent)
-}
