+++ android/festago/app/src/main/java/com/festago/festago/common/analytics/AnalyticsHelper.kt
@@ -1,0 +1,7 @@
+package com.festago.festago.common.analytics
+
+import com.festago.festago.common.analytics.AnalyticsEvent
+
+interface AnalyticsHelper {
+    fun logEvent(event: AnalyticsEvent)
+}
