+++ android/festago/app/src/main/java/com/festago/festago/presentation/ui/studentverification/StudentVerificationViewModel.kt
@@ -2,8 +2,8 @@ package com.festago.festago.presentation.ui.studentverification
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.festago.festago.analytics.AnalyticsHelper
-import com.festago.festago.analytics.logNetworkFailure
+import com.festago.festago.common.analytics.AnalyticsHelper
+import com.festago.festago.common.analytics.logNetworkFailure
 import com.festago.festago.model.StudentVerificationCode
 import com.festago.festago.model.timer.Timer
 import com.festago.festago.model.timer.TimerListener
