+++ android/festago/app/src/test/java/com/festago/festago/presentation/ui/home/ticketlist/TicketListViewModelTest.kt
@@ -1,7 +1,7 @@
 package com.festago.festago.presentation.ui.home.ticketlist
 
 import app.cash.turbine.test
-import com.festago.festago.analytics.AnalyticsHelper
+import com.festago.festago.common.analytics.AnalyticsHelper
 import com.festago.festago.model.Ticket
 import com.festago.festago.presentation.fixture.TicketFixture
 import com.festago.festago.presentation.rule.MainDispatcherRule
