+++ android/festago/app/src/test/java/com/festago/festago/presentation/ui/selectschool/SelectSchoolViewModelTest.kt
@@ -1,7 +1,7 @@
 package com.festago.festago.presentation.ui.selectschool
 
 import app.cash.turbine.test
-import com.festago.festago.analytics.AnalyticsHelper
+import com.festago.festago.common.analytics.AnalyticsHelper
 import com.festago.festago.model.School
 import com.festago.festago.presentation.rule.MainDispatcherRule
 import com.festago.festago.repository.SchoolRepository
