+++ android/festago/app/src/test/java/com/festago/festago/presentation/ui/studentverification/StudentVerificationViewModelTest.kt
@@ -1,7 +1,7 @@
 package com.festago.festago.presentation.ui.studentverification
 
 import app.cash.turbine.test
-import com.festago.festago.analytics.AnalyticsHelper
+import com.festago.festago.common.analytics.AnalyticsHelper
 import com.festago.festago.model.StudentVerificationCode
 import com.festago.festago.presentation.rule.MainDispatcherRule
 import com.festago.festago.repository.SchoolRepository
